fastapi-solution/etl_postgres_to_elastic/pyproject.toml
[tool.poetry]
name = "sprint 3. etl"
version = "0.1.0"
description = "sprint 3 yandex practicum middle python"
authors = ["Sergey Bilibin <s_bilibin2017@yandex.ru.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "==2.9.5"
types-psycopg2 = "==2.9.21.4"
python-dotenv="==0.21.1"
pandas="==1.5.3"
pydantic="==1.10.5"
elasticsearch="==7.13.4"
requests="==2.28.2"
elasticsearch-dsl="==7.4.0"
redis = "==4.5.1"
backoff = "==2.2.1"
SQLAlchemy = "==2.0.5.post1"

[tool.poetry.dev-dependencies]
autopep8 = "^2.0.1"
pytest = "==7.2.1"
black = "==23.1.0"
isort = "==5.12.0"
flake8 = "==6.0.0"
mypy = "==1.0.0"
pre-commit = "==3.0.4"
Flake8-pyproject = "==1.2.2"
pyproject-autoflake = "==1.0.0"



[tool.autoflake]
in-place=true
expand-star-imports=true
remove-duplicate-keys=true
remove-all-unused-imports=true


[tool.black]
line-length = 120
skip-string-normalization=true


[tool.isort]
multi_line_output=3
include_trailing_comma=true
force_grid_wrap=0
use_parentheses=true
line_length=120

[tool.flake8]
ignore = ['E203', 'E266', 'E501', 'W503']
max-line-length = 120
max-complexity = 18
select = ['B','C','E','F','W','T4']
exclude= [
    'app/movies/migrations/*',
    'app/config/*',
    '__pycache__',
    'app/movies/manage.py',
    'settings.py',
    'env',
    '.env'
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
warn_unused_configs = true
disallow_incomplete_defs = true